requirements.txt
# Django
django == 4.2.20
django-cors-headers == 4.7.0
djangorestframework == 3.15.2
djangorestframework_simplejwt == 5.5.0

# Tools
environs == 14.1.1
sentry-sdk == 2.22.0
mysqlclient == 2.2.7
